@@ -2,8 +2,8 @@ package com.blog.DTO;
 
 import com.blog.entity.Comments;
 
-import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.Size;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.Size;
 import java.util.Set;
 
 public class PostBlogDTO {
