@@ -1,6 +1,6 @@
 package com.blog.entity;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Entity
 @Table (name = "comments")
